refactor: split message rendering modules
