+++ web-gui/app/src/runtime/session-events.ts
@@ -1,0 +1,7 @@
+export interface SessionEventEnvelope {
+  id?: string;
+  event_seq?: number;
+  ts?: string;
+  type?: string;
+  payload?: unknown;
+}
